refactor affichage du classement et calcul du score directement sur le devoxxUser

+++ devoxx-game/src/main/java/fr/soat/devoxx/game/forms/UserGameInformation.java
@@ -7,21 +7,16 @@ import java.io.Serializable;
 import java.util.List;
 
 /**
- * Created by IntelliJ IDEA.
- * User: mparisot
- * Date: 12/03/12
- * Time: 18:49
- * To change this template use File | Settings | File Templates.
  */
 public class UserGameInformation implements Serializable {
 
     private static final long serialVersionUID = -1918069635247552059L;
 
     private int currentRanking;
-    private int nbOfPlayers;
+    private long nbOfPlayers;
     private List<UserQuestion> questionsInProgress;
 
-    public UserGameInformation(int currentRanking, int nbOfPlayers, List<UserQuestion> questionsInProgress) {
+    public UserGameInformation(int currentRanking, long nbOfPlayers, List<UserQuestion> questionsInProgress) {
         this.currentRanking = currentRanking;
         this.nbOfPlayers = nbOfPlayers;
         this.questionsInProgress = questionsInProgress;
@@ -31,7 +26,7 @@ public class UserGameInformation implements Serializable {
         return currentRanking;
     }
 
-    public int getNbOfPlayers() {
+    public long getNbOfPlayers() {
         return nbOfPlayers;
     }
